Added checking for data constructors to the type checker.
Three to-do items for the type checker now remain (that I am aware of):

(1) need to implement kind checking
(2) need to implement type checking for case expressions
(3) need to figure out why type variables in the definition signature are not lining up with those in the definition body.

Better error messages would be nice too.